requirements.txt
hume==0.7.5
pillow==11.1.0
python-dotenv==1.0.1
